+++ Neo4jDataImporterToCSharpClasses/CypherWorkbench/CypherWorkbench.cs
@@ -1,0 +1,12 @@
+﻿namespace Neo4jJsonToCSharpClasses.CypherWorkbench;
+
+using Newtonsoft.Json;
+
+public class CypherWorkbench
+{
+    [JsonProperty("metadata")]
+    public CypherWorkbenchMetadata? Metadata { get; set; }
+
+    [JsonProperty("dataModel")]
+    public CypherWorkbenchDataModel? DataModel { get; set; }
+}
