+++ Neo4jDataImporterToCSharpClasses/CypherWorkbench/CypherWorkbenchMetadata.cs
@@ -1,0 +1,9 @@
+﻿namespace Neo4jJsonToCSharpClasses.CypherWorkbench;
+
+using Newtonsoft.Json;
+
+public class CypherWorkbenchMetadata
+{
+    [JsonProperty("cypherWorkbenchVersion")]
+    public Version? Version { get; set; }
+}
